+++ core/analysis/analysis.hpp
@@ -7,3 +7,5 @@
 
 #include "analysis_engine.hpp"
 #include "analysis_model.hpp"
+#include "log_level_counts.hpp"
+#include "log_line_classifier.hpp"